refactor(032/substrate): split oci_pull.rs into mod.rs + reference.rs + platform.rs + tarball.rs
First commit of milestone 032 — restructures the milestone-031
single-file `oci_pull.rs` (488 LOC) into a 4-submodule directory
to set up the oci-client → oci-spec swap that follows in commit 2.

**Behavior is unchanged.** This commit is purely a reorganization;
oci-client is still the substrate and the smoke test still pulls
the same images with the same component output.

Submodules:

* `mod.rs` — public surface preserved (`pull_to_tarball`,
  `ImageArgKind`, `detect_image_arg_kind`, `host_oci_arch`).
  Module-doc-comment updated with substrate-evolution notes.
  `pull_to_tarball` body still calls `oci_client::Client::pull`
  but routes through `tarball::*` and `platform::*` helpers
  rather than inline closures.

* `reference.rs` — NEW. `parse_reference(&str) ->
  Result<ImageReference>` parser written from scratch. Implements
  the 4 grammar rules (registry-vs-repo disambiguation via
  `is_registry_segment`; `library/` prefix on docker.io bare
  names; `latest` tag default; digest-precedence-over-tag).
  Cross-validated against today's `oci_client::Reference`
  parser via 8 table-test cases covering the typical ref
  shapes (`alpine`, `alpine:3.19`, `library/alpine:3.19`,
  `docker.io/library/alpine:3.19`, `gcr.io/foo/bar:tag`,
  `localhost:5000/foo/bar:tag`, digest form, distroless).
  Plus 6 negative cases (empty input, empty tag, empty
  digest, malformed digest, registry-without-repo, type-alias
  for the test-tuple shape per clippy::very-complex-types).
  This parser is NOT YET WIRED IN — `mod.rs` still uses
  `oci_client::Reference`. Used in `detect_image_arg_kind` for
  the ref-validation path. Commit 2 wires it into the pull path.

* `platform.rs` — NEW. `resolve_manifest_list_to_linux(entries,
  target_arch) -> Result<digest>` extracted from today's inline
  closure in `pull_to_tarball`. Decoupled from any specific
  crate's manifest-list type via a borrowed-view
  `ManifestListEntry<'a>` struct, so this works equally well
  with oci-client's `ImageIndexEntry` (today) and oci-spec's
  `Descriptor` (commit 2). 4 tests covering happy path,
  duplicate-platform tie-break, no-match error message, and
  non-linux-OS-entry filtering.

* `tarball.rs` — assembled from today's
  `assemble_docker_save_tarball` / `decompress_layer` /
  `append_tarball_entry` / `sha256_hex` /
  `assert_layers_supported` helpers. Type signatures preserved
  (still take `oci_client::ImageData` / `ImageLayer`); commit
  2 will refactor to take primitive types. Existing tests
  preserved verbatim.

Verification:
* `./scripts/pre-pr.sh` (default profile) clean.
* `cargo +stable test -p mikebom --features oci-registry` clean.
* `cargo +stable clippy --workspace --all-targets --features
  oci-registry` clean (had to fix 2 lints surfaced by the
  reorganization: `clippy::doc_lazy_continuation` on the
  module-doc-comment block; `clippy::very_complex_types` on
  the test fixture's tuple — factored to a `ParseCase` type
  alias).
